BackEnd/requirements.txt
Django==2.0.7
django-cors-headers==2.3.0
django-filter==1.1.0
djangorestframework==3.8.2
factory-boy==2.11.1
Faker==0.8.17
Markdown==2.6.11
Pillow==5.2.0
pkg-resources==0.0.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
text-unidecode==1.2
